Get rid of chainstate macro, still with some repetitions of code
